+++ modules/hanna-react/src/utils/useDomid.ts
@@ -1,4 +1,4 @@
-import * as React from 'react';
+import React from 'react';
 import domid from '@hugsmidjan/qj/domid';
 
 // @ts-expect-error  (transparently feature-detect useId hook, which is introduced in React@18)